Add files via upload
Development Process:
DB Browser for SQLite for Backend: We used DB Browser for SQLite for Database Management and Adding Questions
C++ for integration: QML and DB Browser for SQLite cannot be integrated directly.So C++ is used as an integrating language
QML for Frontend: QML is an Modelling Language.  It's particularly suited for designing interactive and dynamic UIs. It is also used for  highly performant application.
 Testing and Quality Assurance: 
QT Creator 10.0.1(Community) version is used in our project
DB Browser for SQLite for Database
Figma For UI Design
Desktop Qt 5.15.2 MinGW 64-bit for debugging